Create a proper solution (with funky structure)
Might change it later, I am still trying to figure out the nicest
way of keeping the source files. Anyway. I run AoC2015.cpp - this
is the main(). From there, we call the input file and based on
that, we call the source files in the root. I wanted to do it
without the header files, but it is not possible (without putting
everything in a single source file) - hence the headers.

Also I am not sure if I need all the project files source controlled.