[day 08] improve second part using `repr`
I tried `repr` initially, but I wasn't getting correct results in all
cases. This is because `repr` encloses a string with single quotes `'`,
and only escapes backslashes. Double quotes `"` do not get escaped.

A workaround was to replace all quotes with backslashes. In that case
`repr` inserts extra backslashes. Since we are interested in the count
only, not in the actual encodig, this does the trick.