[day18] part 1. we have a fixed mesh!
Initially I implemented it with mesh expanding (like in 2020 day 17), so
I needed to put a limit on the mesh dimensions when defining neighbors.

It is pretty slow, I don't know what could be improved here, probably
with a fixed mesh we can generate all the neighbors at once and maybe
put that in a dictionary? Not sure if it would help.