[day24] both parts, assuming rest is splittable
inspired by Reddit solution megathread, I implemented this where we
assume (based on the input data) that once we find the smallest group,
the rest will be splittable into remaining groups.

the fact that we use itertools.permutations on a sorted list makes sure
that whatever the first found smallest group is, it satisfies the
"smallest cosmic entanglement" condition, ie. that this is not only the
smallest group size, but the elements in the group are also smallest
possible.

to do: add check which proves that the remaining numbers WILL split into
equal sum groups.