[day07] use `eval` with `join` to evaluate expr.
The hacky abs at the return is here because of the way the bitwise
complement is defined in order to work with this method of joining.

It says in the problem that the maximum value is 65535 and that wires
only take positive values. I figured out that all logical operators do
exactly what they are supposed to do in the puzzle description, except
for the ^ (not) operator. For example, ^n will return -n - 1... So I
figured, ok we actually need 65535 - x. Since the operator is joined
after the wire it works on, the only way to fix this is to actually use
x - 65535 and return the absolute value. abs will not affect any other
operation as they all return positive values by default.