[day07] add memoization - it runs instantly now
I am still very confused about how decorators work. I understand how
memoization works and I understand that it could be written in a more
generic way (using *args), but since one of the args is a dictionary it
cannot be used as a dict key (and we use dict to store function values).

basically what this does is that it stores all previously calculated
values from the solve_wires function, thus speeding up the recursion (we
do not have to recalculate the same thing over and over again).

I will definitely try solving this again without recursion and
memoization because I *feel* it is possibele (see my cpp solution, ugh),
but I wanted to do it this way to learn something new.

BTW can we think of solve_wires as some sort of depth first search??